fix smoketest cleanup on early abort

+++ smoketest.sh
@@ -12,15 +12,14 @@ set -e
 export DSTASK_GIT_REPO=$(mktemp -d)
 export DSTASK_STATE_FILE=$(mktemp -u)
 export DSTASK_FAKE_PTY=1
-
-UPSTREAM_BARE_REPO=$(mktemp -d)
+export UPSTREAM_BARE_REPO=$(mktemp -d)
 
 cleanup() {
     set +x
     set +e
     rm -rf $DSTASK_GIT_REPO
     rm -rf $UPSTREAM_BARE_REPO
-    rm $DSTASK_STATE_FILE
+    [ -f $DSTASK_STATE_FILE ] && rm $DSTASK_STATE_FILE
     rm dstask
 }
